+++ src/app/lib/main.dart
@@ -1,7 +1,9 @@
 import 'package:app/access/dad_jokes/dad_jokes_repository.dart';
 import 'package:app/access/dad_jokes/favorite_dad_jokes_repository.dart';
+import 'package:app/access/diagnostics/diagnostics_repository.dart';
 import 'package:app/app.dart';
 import 'package:app/business/dad_jokes/dad_jokes_service.dart';
+import 'package:app/business/diagnostics/diagnostics_service.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
@@ -26,6 +28,7 @@ void _registerRepositories() {
     ),
   );
   GetIt.I.registerSingleton(FavoriteDadJokesRepository());
+  GetIt.I.registerSingleton(DiagnosticsRepository());
 }
 
 /// Registers the services.
@@ -34,6 +37,11 @@ void _registerServices() {
     DadJokesService(
       GetIt.I.get<DadJokesRepository>(),
       GetIt.I.get<FavoriteDadJokesRepository>(),
+    ),
+  );
+  GetIt.I.registerSingleton(
+    DiagnosticsService(
+      GetIt.I.get<DiagnosticsRepository>(),
     ),
   );
 }