+++ CHANGELOG.md
@@ -21,3 +21,4 @@ Prefix your items with `(Template)` if the change is about the template and not 
 - Fixed PowerShell script that was copying source control.
 - Added Navigation to the app.
 - Fixed Mergify.
+- Added Diagnostic overlay to the app.